Move tests from android->unit (no android dependencies)
